Change cache handling
Update cache to be used indefinitely regardless of staleness.
Add cache clear command.
Change api retrieval to error when a pokemon, type, move, or ability is
not present in a generation.
Change the --game option to default to the latest game instead of
hardcoding it.
Optimize move and game data retrieval by using FuturesUnordered instead.
Refactor converting game-to-generation to be non-async.